requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==3.0.3
dnspython==2.7.0
Flask==3.0.3
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.5
MarkupSafe==3.0.2
narwhals==1.35.0
nest-asyncio==1.6.0
packaging==24.2
plotly==6.0.1
py4j==0.10.9.7
pymongo==4.12.0
pyspark==3.5.5
requests==2.32.3
retrying==1.3.4
setuptools==78.1.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==3.0.3
dnspython==2.7.0
Flask==3.0.3
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.5
MarkupSafe==3.0.2
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
plotly==6.0.1
py4j==0.10.9.7
pymongo==4.12.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
setuptools==78.1.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
